% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coversion_id_genes.R
\name{conversion_id_genes}
\alias{conversion_id_genes}
\title{Gene ID Conversion}
\usage{
conversion_id_genes(list_ensembl, keyid, outputid)
}
\arguments{
\item{list_ensembl}{A list containing the values to filter with.}

\item{keyid}{The column in the dataframe to use for filtering.}

\item{outputid}{A vector of column names to select.}
}
\value{
A dataframe containing the filtered and selected results.
}
\description{
This function filters an SQLite dataframe based on a list of values and selects specific columns.
}
