% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_id_genes_mouse.R
\name{conversion_id_genes_mouse}
\alias{conversion_id_genes_mouse}
\title{Convert Gene IDs using Mouse Database}
\usage{
conversion_id_genes_mouse(list_ensembl, keyid, outputid)
}
\arguments{
\item{list_ensembl}{A vector of Ensembl gene IDs to filter.}

\item{keyid}{The column in the dataframe to use for filtering (e.g., "ENSEMBL").}

\item{outputid}{A vector of column names to return.}
}
\value{
A dataframe containing the filtered and selected results.
}
\description{
This function filters an SQLite dataframe for mouse genes based on a list of values and selects specific columns.
}
\details{
This function connects to the SQLite database specific for mouse genes (\code{mgeneslator.sqlite})
stored in the \code{extdata} directory of the package. It checks for the table existence and
performs filtering and selection based on the provided parameters.
}
\examples{
\dontrun{
# Example usage
list_ensembl <- c("ENSMUSG00000064337", "ENSMUSG00000029580")
keyid <- "ENSEMBL"
outputid <- c("SYMBOL", "ENTREZID")

result <- conversion_id_genes_mouse(list_ensembl, keyid, outputid)
print(result)
}
}
